Factor out model fixtures, adjust tests
