+++ test/unit/data/model/mapping/test_tsi_model_mapping.py
@@ -14,6 +14,7 @@ from .testing_utils import (
     AbstractBaseTest,
     collection_consists_of_objects,
 )
+from .testing_utils.tsi_model_fixtures import *
 from ..common import (
     dbcleanup,
     dbcleanup_wrapper,
@@ -286,65 +287,3 @@ class TestToolVersionAssociation(BaseTest):
             assert stored_obj.parent_id == parent_tool_version.id
 
         delete_from_database(session, [tool_version, parent_tool_version])
-
-
-# Misc. helper fixtures.
-
-
-@pytest.fixture(scope="module")
-def init_model(engine):
-    model.mapper_registry.metadata.create_all(engine)
-
-
-# Fixtures yielding persisted instances of models, deleted from the database on test exit.
-
-
-@pytest.fixture
-def repository(session):
-    instance = model.ToolShedRepository()
-    yield from dbcleanup_wrapper(session, instance)
-
-
-@pytest.fixture
-def repository_repository_dependency_association(session):
-    instance = model.RepositoryRepositoryDependencyAssociation()
-    yield from dbcleanup_wrapper(session, instance)
-
-
-@pytest.fixture
-def repository_dependency(session, repository):
-    instance = model.RepositoryDependency(repository.id)
-    yield from dbcleanup_wrapper(session, instance)
-
-
-@pytest.fixture
-def tool_dependency(session, repository):
-    instance = model.ToolDependency()
-    instance.tool_shed_repository = repository
-    instance.status = "a"
-    yield from dbcleanup_wrapper(session, instance)
-
-
-@pytest.fixture
-def tool_version(session):
-    instance = model.ToolVersion()
-    yield from dbcleanup_wrapper(session, instance)
-
-
-# Fixtures yielding factory functions.
-
-
-@pytest.fixture
-def tool_version_association_factory():
-    def make_instance(*args, **kwds):
-        return model.ToolVersionAssociation(*args, **kwds)
-
-    return make_instance
-
-
-@pytest.fixture
-def tool_version_factory():
-    def make_instance(*args, **kwds):
-        return model.ToolVersion(*args, **kwds)
-
-    return make_instance
